Fix inexplicable packet malformed read and minor nullpointer crash
